requirements.txt
# Weaviate client for vector database operations
weaviate-client==4.16.10

# Environment variable management
python-dotenv==1.0.0

# Data validation and modeling
pydantic>=2.8.0
pydantic-settings>=2.10.0

# Beautiful terminal output and progress bars
rich==13.7.0

# Python client for Ollama
ollama==0.5.4

# JSON handling and date parsing utilities
python-dateutil==2.8.2

# HTTP requests (dependency for weaviate-client)
requests==2.31.0

# Type hints support
typing-extensions>=4.12.2

# FastAPI for RAG API server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9  # Required for file upload endpoints

# Additional providers support
openai>=1.0.0  # For OpenAI embeddings and generation
httpx>=0.24.0  # For OpenRouter API calls